Fixed a couple bugs having to do with prefetches getting stalled by demand fetches. Related bugs also manifested in prefetches getting out of order; fixed those too.

+++ library_project/src/edu/umich/eac/CacheFetcher.java
@@ -15,6 +15,6 @@ public abstract class CacheFetcher<V> {
                                     int worstBandwidthUp,
                                     int worstRTT) {
         // default implementation; assume downstream transfer dominates
-        return bytesToTransfer() / worstBandwidthDown;
+        return ((double) bytesToTransfer()) / ((double) worstBandwidthDown);
     }
 }
